backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
sqlalchemy[asyncio]
asyncpg
pyserial
reportlab
psycopg2-binary
